requirements.txt
discord.py
python-dotenv
PyNaCl
pymupdf
tqdm
sentence-transformers
numpy
google-genai